fix: change tokenId type from Long to Integer in Token class

+++ src/main/java/com/mongs/book_social_network/user/Token.java
@@ -15,7 +15,7 @@ import java.time.LocalDateTime;
 public class Token {
     @Id
     @GeneratedValue
-    private Long tokenId;
+    private Integer tokenId;
 
     private String token;
     private LocalDateTime createdAt;
